fixed active state for keyboards
