+++ bloom_site/mix.exs
@@ -54,7 +54,7 @@ defmodule BloomSite.MixProject do
       {:gettext, "~> 0.20"},
       {:jason, "~> 1.2"},
       {:plug_cowboy, "~> 2.5"},
-      {:bloom, "~> 0.0.7"},
+      {:bloom, path: "../"},
       # {:bloom, path: "..", override: true},
       {:phoenix_storybook, "~> 0.6.0"}
     ]
